fix(tokens): restore deprecated foundation/background/outline color tokens
Some apps may still use an old version of reset CSS in production before we bump the aliased version
of @warp-ds/css. That stylesheet relies on old color tokens that were removed in favour of semantic
tokens. In order for layouts to use the new tokens and then let apps safely align with the other
dependencies and stylesheets, we need to enure backward compatibility of the tokens.