fix(alert): override font-size/line-height of h3 title in Alert (#64)
Alert title should be a heading to ensure accessibility. In @warp-ds/vue Alert accepts a title prop, which will now be a h3 instead of a p with bold font-weight, so in order for that to have the same size as p, we need to assign that heading element a 'text-s' class. A corresponding @warp-ds/vue update is required for this change to be applied.