fix(expandable): fix long titles overflowing chevron (#89)
* fix: expandables with long titles could overflow chevron

* fix: set margin-left on chevronNonBox

Now that we don't apply absolute position on chevronBox, chevronNonBox can be changed to ml-8. This
also fixes wrong alignment of chevron in Elements

* fix: set w-full on expandable.title

In React and Elements we are wrapping text and chevron in a div, which then holds expandable.title
classes. In order for the chevron to be positioned at the end of a box Expandable there, we need to
apply w-full here.

---------

Co-authored-by: BalbinaK <balbuhhha@gmail.com>